chore: move tests around
