+++ packages/ws/package.json
@@ -69,6 +69,7 @@
 		"@types/ws": "^8.5.3",
 		"c8": "^7.11.3",
 		"eslint": "^8.19.0",
+		"mock-socket": "^9.1.5",
 		"prettier": "^2.7.1",
 		"tsup": "^6.1.3",
 		"typescript": "^4.7.4",
